Ajuste Semillas|Migraciones|Jsons|Modelos

+++ database/migrations/2024_08_21_000542_create_usuario_table.php
@@ -25,10 +25,10 @@ class CreateUsuarioTable extends Migration
             $table->unsignedBigInteger('cod_ips');
             $table->foreign('cod_ips')->references('cod_ips')->on('departamentoxips');
 
-            $table->integer('cod_documento');
+            $table->unsignedBigInteger('cod_documento');
             $table->foreign('cod_documento')->references('cod_documento')->on('tipo_de_documento');
 
-            $table->integer('cod_poblacion');
+            $table->unsignedBigInteger('cod_poblacion');
             $table->foreign('cod_poblacion')->references('cod_poblacion')->on('poblacion_diferencial');
 
             $table->date('fec_diag_usuario');
